Refactor ui_manager.py into modular components (1140→167 lines)
- Split large ui_manager.py into specialized components for better maintainability
- Created modular architecture with WidgetFactory, EventHandlers, WeatherDisplayManager, etc.
- Maintained backward compatibility and all tests passing
- Reduced file length by 85% while improving separation of concerns

Components created:
- alerts_manager.py: Alerts display and management logic
- data_extractors.py: Data extraction for taskbar icons
- data_formatters.py: Weather data formatting for display
- display_managers.py: Weather display management
- event_handlers.py: UI event binding and handling
- ui_utils.py: Utility functions for UI operations
- weather_source_manager.py: Weather source detection and UI updates
- widget_factory.py: UI widget creation and layout
